Add Dockerfile for all the projects and the SDL
A directory is added with files to build a Docker image containing the
projects [Clog](https://github.com/Moonstroke/Clog), a simple logging
system in C, and [CUTE](https://github.com/Moonstroke/CUTE), a unit
testing engine in C, and
[C-Arrays](https://github.com/Moonstroke/C-Arrays), a repo defining
several useful data structures for plain C.

The [SDL](https://libsdl.org) is a free and open-source low-level
multimedia library. It is mainly used to develop games (for example
Half-Life, Broken Sword).

Simply run:
    docker build clog+cute+c-arrays+sdl

from the root of the repository to build the image.