Add Dockerfile for the projects Clog, CUTE & C-Arrays
A directory is added with files to build a Docker image containing the
projects [Clog](https://github.com/Moonstroke/Clog), a simple logging
system in C, and [CUTE](https://github.com/Moonstroke/CUTE), a unit
testing engine in C, and
[C-Arrays](https://github.com/Moonstroke/C-Arrays), a repo defining
several useful data structures for plain C.

In the same way that one cannot have CUTE without Clog, C-Arrays cannot
be used on its own, because it uses the other two projects.

Simply run:
    docker build clog+cute+c-arrays

from the root of the repository to build the image.