Add Dockerfile for the projects Clog & CUTE
A directory is added with files to build a Docker image containing the
projects [Clog](https://github.com/Moonstroke/Clog), a simple logging
system in C, and [CUTE](https://github.com/Moonstroke/CUTE), which is a
unit testing engine in C.

CUTE cannot be used on its own because it uses the other project to
work.

Simply run:
    docker build clog+cute

from the root of the repository to build the image.